requirements.txt
pyTelegramBotAPI
pytesseract
python-dotenv
emoji~=1.6.3
